Add thread_rng

+++ codes/day_73_rand/src/main.rs
@@ -3,7 +3,7 @@ use rand::Rng;
 fn main() {
 
     let mut rng = rand::thread_rng();
-    let x: f64 = rng.gen();
+    let x: i32 = rng.gen();
     println!("{}", x);
     println!("{:?}", rng.gen::<(f64, bool)>());    
 }
